chore: refactor in multiple modules
